+++ docs/source/conf.py
@@ -58,7 +58,7 @@ author = u'Matthew Hartley and Tjelvar Olsson'
 # built documents.
 #
 # The short X.Y version.
-version = u'0.8.0'
+version = u'0.9.0'
 # The full version, including alpha/beta/rc tags.
 release = version